230718_2106_Change variable types(float->double)

+++ mimic/include/fragment.hh
@@ -5,13 +5,13 @@ typedef int color_t;
 struct Fragment
 {
     int x, y;
-    double z;
-    double*  f_data; // interpolated input data
+    float z;
+    float*  f_data; // interpolated input data
 };
 
 struct ShadedFragment
 {
     int x, y;
-    double z;
+    float z;
     color_t color;
 };
